Final bit of tidying the bytez impl.
