+++ src/providers/bytez/utils.ts
@@ -1,3 +1,6 @@
+import { ParameterConfig } from '../types';
+import { BytezInferenceChatCompleteConfig } from './chatComplete';
+
 class LRUCache<K, V> {
   private size: number;
   private map: Map<K, V>;
@@ -52,4 +55,75 @@ class LRUCache<K, V> {
   }
 }
 
-export { LRUCache };
+function bodyAdapter(requestBody: Record<string, any>) {
+  for (const [param, paramConfig] of Object.entries(
+    BytezInferenceChatCompleteConfig
+  )) {
+    const hasParam = Boolean(requestBody[param]);
+
+    // first assign defaults
+    if (!hasParam) {
+      const { default: defaultValue, required } =
+        paramConfig as ParameterConfig;
+
+      // if it's required, throw
+      if (required) {
+        throw new Error(`Param ${param} is required`);
+      }
+
+      // assign the default value
+      if (defaultValue !== undefined && requestBody[param] === undefined) {
+        requestBody[param] = defaultValue;
+      }
+    }
+  }
+
+  // now we remap everything that has an alias, i.e. "prop" on propConfig
+  for (const key of Object.keys(requestBody)) {
+    const paramObj = BytezInferenceChatCompleteConfig[key] as
+      | ParameterConfig
+      | undefined;
+
+    if (paramObj) {
+      const { param: alias } = paramObj;
+
+      if (key !== alias) {
+        requestBody[alias] = requestBody[key];
+        delete requestBody[key];
+      }
+    }
+  }
+
+  // now we adapt to the bytez input signature
+  // props to skip
+  const skipProps: Record<string, boolean> = {
+    model: true,
+  };
+
+  // props that cannot be removed from the body
+  const reservedProps: Record<string, boolean> = {
+    stream: true,
+    messages: true,
+  };
+
+  const adaptedBody: Record<string, any> = { params: {} };
+
+  for (const [key, value] of Object.entries(requestBody)) {
+    // things like "model"
+    if (skipProps[key]) {
+      continue;
+    }
+
+    // things like "messages", "stream"
+    if (reservedProps[key]) {
+      adaptedBody[key] = value;
+      continue;
+    }
+    // anything else, e.g. max_new_tokens
+    adaptedBody.params[key] = value;
+  }
+
+  return adaptedBody;
+}
+
+export { LRUCache, bodyAdapter };
